@@ -7,7 +7,7 @@
     "download",
     "distube"
   ],
-  "version": "4.16.8",
+  "version": "4.16.9",
   "repository": {
     "type": "git",
     "url": "git://github.com/distubejs/ytdl-core.git"
@@ -30,18 +30,18 @@
     "typings"
   ],
   "dependencies": {
-    "http-cookie-agent": "^6.0.8",
+    "http-cookie-agent": "^7.0.1",
     "https-proxy-agent": "^7.0.6",
     "m3u8stream": "^0.8.6",
     "miniget": "^4.2.3",
     "sax": "^1.4.1",
-    "tough-cookie": "^5.1.0",
-    "undici": "^7.3.0"
+    "tough-cookie": "^5.1.2",
+    "undici": "^7.8.0"
   },
   "devDependencies": {
-    "@types/node": "^22.12.0",
-    "prettier": "^3.4.2",
-    "typescript": "^5.7.3"
+    "@types/node": "^22.15.2",
+    "prettier": "^3.5.3",
+    "typescript": "^5.8.3"
   },
   "engines": {
     "node": ">=20.18.1"
