@@ -1,6 +1,6 @@
 {
   "name": "http-cookie-agent",
-  "version": "6.0.8",
+  "version": "7.0.1",
   "description": "Allows cookies with every Node.js HTTP clients.",
   "keywords": [
     "agent",
@@ -32,7 +32,8 @@
   "author": "3846masa <3846masahiro+git@gmail.com>",
   "exports": {
     "./http": "./http/index.js",
-    "./undici": "./undici/index.js"
+    "./undici": "./undici/index.js",
+    "./undici/v6": "./undici/v6/index.js"
   },
   "files": [
     "dist",
@@ -60,27 +61,27 @@
     "agent-base": "^7.1.3"
   },
   "devDependencies": {
-    "@3846masa/configs": "github:3846masa/configs#8ad26d06fa6071bac299dcc44635329adf69e1e0",
-    "@babel/cli": "7.26.4",
-    "@babel/core": "7.26.0",
+    "@3846masa/configs": "github:3846masa/configs#e259de36675e7ccd48a0285d325cb9be710f4370",
+    "@babel/cli": "7.27.0",
+    "@babel/core": "7.26.10",
     "@babel/plugin-proposal-explicit-resource-management": "7.25.9",
-    "@babel/preset-env": "7.26.0",
-    "@babel/preset-typescript": "7.26.0",
+    "@babel/preset-env": "7.26.9",
+    "@babel/preset-typescript": "7.27.0",
     "@hapi/wreck": "18.1.0",
     "@jest/globals": "29.7.0",
     "@semantic-release/changelog": "6.0.3",
     "@semantic-release/git": "10.0.1",
     "@types/babel__core": "7.20.5",
     "@types/needle": "3.3.0",
-    "@types/node": "18.19.68",
+    "@types/node": "20.17.30",
     "@types/request": "2.48.12",
-    "@types/semver": "7.5.8",
+    "@types/semver": "7.7.0",
     "@types/superagent": "8.1.9",
-    "agentkeepalive": "4.5.0",
-    "axios": "1.7.9",
+    "agentkeepalive": "4.6.0",
+    "axios": "1.8.4",
     "babel-jest": "29.7.0",
-    "disposablestack": "1.1.6",
-    "got": "12.6.1",
+    "disposablestack": "1.1.7",
+    "got": "14.4.7",
     "http-proxy-agent": "7.0.2",
     "jest": "29.7.0",
     "needle": "3.3.1",
@@ -88,27 +89,28 @@
     "phin": "3.7.1",
     "proxy": "2.2.0",
     "request": "2.88.2",
-    "rimraf": "5.0.10",
-    "semantic-release": "22.0.12",
-    "semver": "7.6.3",
-    "superagent": "10.1.1",
-    "tough-cookie": "5.0.0",
-    "typescript": "5.7.2",
-    "undici": "6.21.0",
-    "urllib": "4.4.0"
+    "rimraf": "6.0.1",
+    "semantic-release": "24.2.3",
+    "semver": "7.7.1",
+    "superagent": "10.2.0",
+    "tough-cookie": "5.1.2",
+    "typescript": "5.8.3",
+    "undici": "7.8.0",
+    "undici__v6": "npm:undici@6.21.2",
+    "urllib": "4.6.11"
   },
   "peerDependencies": {
     "tough-cookie": "^4.0.0 || ^5.0.0",
-    "undici": "^5.11.0 || ^6.0.0"
+    "undici": "^7.0.0"
   },
   "peerDependenciesMeta": {
     "undici": {
       "optional": true
     }
   },
-  "packageManager": "pnpm@9.15.0",
+  "packageManager": "pnpm@10.8.0",
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=20.0.0"
   },
   "pnpm": {
     "overrides": {
