@@ -105,43 +107,30 @@
     "prepare": "husky && node ./scripts/platform-shell.js"
   },
   "devDependencies": {
-    "@fastify/busboy": "2.1.1",
-    "@matteo.collina/tspl": "^0.1.1",
-    "@sinonjs/fake-timers": "^11.1.0",
-    "@types/node": "~18.19.50",
+    "@fastify/busboy": "3.1.1",
+    "@matteo.collina/tspl": "^0.2.0",
+    "@sinonjs/fake-timers": "^12.0.0",
+    "@types/node": "^18.19.50",
     "abort-controller": "^3.0.0",
-    "borp": "^0.15.0",
+    "borp": "^0.19.0",
     "c8": "^10.0.0",
     "cross-env": "^7.0.3",
     "dns-packet": "^5.4.0",
+    "esbuild": "^0.25.2",
+    "eslint": "^9.9.0",
     "fast-check": "^3.17.1",
-    "form-data": "^4.0.0",
-    "formdata-node": "^6.0.3",
     "https-pem": "^3.0.0",
     "husky": "^9.0.7",
     "jest": "^29.0.2",
-    "jsdom": "^24.0.0",
+    "neostandard": "^0.12.0",
     "node-forge": "^1.3.1",
-    "pre-commit": "^1.2.2",
     "proxy": "^2.1.1",
-    "snazzy": "^9.0.0",
-    "standard": "^17.0.0",
-    "tsd": "^0.31.0",
-    "typescript": "^5.0.2",
+    "tsd": "^0.31.2",
+    "typescript": "^5.6.2",
     "ws": "^8.11.0"
   },
   "engines": {
-    "node": ">=18.17"
-  },
-  "standard": {
-    "env": [
-      "jest"
-    ],
-    "ignore": [
-      "lib/llhttp/constants.js",
-      "lib/llhttp/utils.js",
-      "test/fixtures/wpt"
-    ]
+    "node": ">=20.18.1"
   },
   "tsd": {
     "directory": "test/types",
